Fix a crash when loading snapshot due to type mismatch commit_hash:e8fcc50ac084676d3abf377ff457af7ea0340756

+++ yt/yt/server/lib/transaction_supervisor/transaction_supervisor.cpp
@@ -3078,7 +3078,7 @@ private:
 
         // COMPAT(h0pless): Remove after 26.2.
         if (static_cast<ETransactionSupervisorReign>(context.GetVersion()) < ETransactionSupervisorReign::StrongOrderingTags) {
-            int nextStronglyOrderedTransactionSequenceNumber;
+            i64 nextStronglyOrderedTransactionSequenceNumber;
             std::map<TTimestamp, int> preparedTransactionsTimestamps;
             std::set<i64> uncommittedTransactionSequenceNumbers;
             THashMap<TTransactionId, i64> transactionIdToSequenceNumber;
